Support negative literal patterns in match expressions
Add support for patterns like `-1 => ...` in match arms. Previously,
only positive integer literals were supported as patterns.

Implementation:
- Add Pattern::NegInt variant to AST for negative integer patterns
- Parse unary minus followed by integer literal as a negative pattern
- Change pattern integer types from u64 to i64 throughout the IR stack
  (RIR, AIR, CFG) to handle signed values correctly
- Update both x86_64 and aarch64 backends

Fixes rue-b2n

+++ crates/rue-codegen/src/aarch64/cfg_lower.rs
@@ -2646,12 +2646,11 @@ impl<'a> CfgLower<'a> {
 
                 // Generate comparison and jump for each case
                 for (value, target) in cases {
-                    // Compare scrutinee with case value
-                    // Cast to i64 to preserve bit pattern
+                    // Compare scrutinee with case value (supports signed values for negative patterns)
                     let imm_vreg = self.mir.alloc_vreg();
                     self.mir.push(Aarch64Inst::MovImm {
                         dst: Operand::Virtual(imm_vreg),
-                        imm: *value as i64,
+                        imm: *value,
                     });
                     self.mir.push(Aarch64Inst::CmpRR {
                         src1: Operand::Virtual(scrutinee_vreg),
